+++ NEWS.md
@@ -1,6 +1,7 @@
 # gmvjoint version 0.2.0
 * Dynamic predictions added
-* ROC function added
+* ROC/AUC function added
+* Function to generate bootstrapped confidence intervals for AUC added.
 * Various S3 methods for joint objects added.
 
 # gmvjoint version 0.1.0